Move classes and structs to separate files
